+++ extras/test/cores/stm32/Print.h
@@ -12,6 +12,7 @@
 struct Print {
   virtual size_t write(const uint8_t *buffer, size_t size) = 0;
   virtual size_t write(uint8_t data) = 0;
+  virtual void flush() {}
 
   virtual int availableForWrite() {
     return 0;