+++ src/StreamUtils/Configuration.hpp
@@ -8,7 +8,8 @@
 #if defined(ARDUINO_ARCH_ESP8266) || defined(ARDUINO_ARCH_SAMD) ||    \
     defined(ARDUINO_ARCH_AVR) ||                                      \
     (defined(ARDUINO_ARCH_ESP32) && ESP_ARDUINO_VERSION_MAJOR >= 2 && \
-     ESP_ARDUINO_VERSION_PATCH >= 3)
+     ESP_ARDUINO_VERSION_PATCH >= 3) ||                               \
+    (defined(ARDUINO_ARCH_STM32) && STM32_CORE_VERSION_MAJOR >= 2)
 #define STREAMUTILS_PRINT_FLUSH_EXISTS 1
 #else
 #define STREAMUTILS_PRINT_FLUSH_EXISTS 0