+++ CHANGELOG.md
@@ -4,7 +4,7 @@ StreamUtils - Change log
 HEAD
 ----
 
-* Add support for `Print::flush()` on ESP32
+* Add support for `Print::flush()` on ESP32 and STM32
 
 1.6.3 (2022/05/11)
 -----