+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Add support for `Print::flush()` on ESP32 and STM32
+* Add `ProgmemStream`
 
 1.6.3 (2022/05/11)
 -----